+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Timeout configuration for `HttpClient` ([#163](https://github.com/stac-utils/stac-asset/pull/163))
 - **pytest-recording** and vcr marks to a couple tests ([#166](https://github.com/stac-utils/stac-asset/pull/166))
 
 ## [0.3.1] - 2024-05-13